requirements.txt
# AI Log Analyzer — Python Dependencies
# Install all at once with: pip install -r requirements.txt

# Web framework — handles routing, requests, and HTML rendering
flask==3.0.2

# Loads environment variables from .env file into os.environ
python-dotenv==1.0.1

# NOTE: No extra AI library needed!
# We call the Gemini API directly using Python's built-in urllib.
# This means fewer things to install and fewer things to break.
